Initial revision of a ViE fuzz test. The idea is to inject randomized RTP packets and see what the video engine does.
There are some small refactorings in here, but the real focus of this CL is in vie_autotest_rtp_fuzz.cc. This patch is mostly here to get a discussion going.

On my initial test the video engine doesn't recover, at least within 10 seconds of running with untampered packets. Not sure if this is according to specification though.

Ideas:
  - Generate random packets with correct RTP headers to get further into the code.
  - Don't generate fresh random data, but rather corrupt bits here and there in small amounts.

BUG=
TEST=

Review URL: https://webrtc-codereview.appspot.com/383001

git-svn-id: http://webrtc.googlecode.com/svn/trunk@1714 4adac7df-926f-26a2-2b94-8c16560cd09d