Based on Chromium's script at http://src.chromium.org/viewvc/chrome/trunk/src/tools/coverity/coverity.py?view=markup
Requires Coverity Static Analysis to be installed on the machine running the script.
The password is kept in a local file that is read by the script.

Not tested on Windows but Chromium seem to run their script on Windows. We'll use Linux for the periodic execution.

It has been running without issues for a month now.

BUG=
TEST=Tested successfully on my local Linux workstation with tools/coverity/coverity.py --source-dir=$HOME/code/webrtc/trunk --coverity-bin-dir=$HOME/coverity-integrity-center/static-analysis/bin --coverity-intermediate-dir=$HOME/cov-temp --coverity-db-host=webrtc-coverity.eem.corp.google.com --coverity-target=Linux --coverity-password-file=$HOME/coverity-password-file

Review URL: https://webrtc-codereview.appspot.com/345001

git-svn-id: http://webrtc.googlecode.com/svn/trunk@1760 4adac7df-926f-26a2-2b94-8c16560cd09d