Set working dir for test run script + update resources
By changing the working directory for the executing script to the same
directory as the script is located in, it is possible to run the script
standing in a higher-level directory (otherwise the input file relative
paths become invalid).

This CL also changes the input file path for the audio_e2e_test test to
assume the file is located resources.

BUG=none
TEST=locally executed the tests standing in trunk/

Review URL: https://webrtc-codereview.appspot.com/1061009

git-svn-id: http://webrtc.googlecode.com/svn/trunk@3422 4adac7df-926f-26a2-2b94-8c16560cd09d