Fix frame_editing_unittest reference file handling.
This test was initializing strings for reference video files in a static
context, which makes is against the style guide and also makes the paths
become invalid when the test is launched from a working directory
outside the checkout.

Moving the initialization into the test fixture solves this.

BUG=none
TEST=Local execution launched from a directory outside the checkout tree on Win, Mac and Linux + trybots (for
compilation as they don't yet run the tools_unittests).

Review URL: https://webrtc-codereview.appspot.com/1178004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@3647 4adac7df-926f-26a2-2b94-8c16560cd09d