This is related to an earlier CL of enabling Opus 48 kHz. https://webrtc-codereview.appspot.com/16619005/
It was reverted due to a build bot error, which this CL is to fix. The problem was that when audio conference mixer receives audio frames all at 48 kHz and mixed them, it uses Audio Processing Module (APM) to do a post-processing. However the APM cannot handle 48 kHz input. The current solution is not to allow the mixer to output 48 kHz.

TEST=locally solved https://webrtc-codereview.appspot.com/16619005/

BUG=
R=andrew@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/20779004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@6730 4adac7df-926f-26a2-2b94-8c16560cd09d