PeerConnectionClient.createPeerConnectionClient was calling new PeerConnectionParameters and PeerConnectionClient.createPeerConnectionFactory, .createPeerConnection with invalid arguments.
This CL makes sure the project compiles, it does not ensure the parameters now used are correct!

There may be something strange going on with the build files. I was previously able to recompile the whole project despite of the incorrect code, until I changed the file and tried again.
The changes made are just so that it will compile. The code should likely be updated by someone who knows what he/she is doing.

TBR=glaznev@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/45369004

Cr-Commit-Position: refs/heads/master@{#8526}
git-svn-id: http://webrtc.googlecode.com/svn/trunk@8526 4adac7df-926f-26a2-2b94-8c16560cd09d