Exclude legacy objc API tests properly.
In https://codereview.webrtc.org/1691463002 the legacy objc tests
were moved into a new GYP file, which was conditioned so it's not
included for platforms that cannot build it. This condition contained
an error which makes the GYP file being processed even if include_tests=0,
which causes errors in downstream code.

BUG=webrtc:5419
NOTRY=True

Review URL: https://codereview.webrtc.org/1701053005

Cr-Commit-Position: refs/heads/master@{#11650}