Changed the semantics of Buffer::Clear to not alter the capacity
Also added a test for Clear to ensure this invariant holds.

With this change, it is easy to empty a Buffer and reuse its storage. Further down the line, code filling data into a Buffer could be written to just append to it, with the caller determining if the Buffer should first be cleared or not.

There is currently only one use of Buffer::Clear (in AudioEncoderCopyRed::Reset()) and it should benefit from the change, by not requiring a reallocation after Reset.

Review URL: https://codereview.webrtc.org/1707693002

Cr-Commit-Position: refs/heads/master@{#11680}