VoE: Handle empty playout timestamp differently
With this change, the VoE Channel will handle the case of an empty
playout timestamp (from audio_coding_->PlayoutTimestamp())
differently. The purpose of the change is to prepare for an upcoming
change in NetEq where empty values will be returned more often (i.e.,
not only before the first packet is received).

BUG=webrtc:5669

Review URL: https://codereview.webrtc.org/1857183002

Cr-Commit-Position: refs/heads/master@{#12261}