Changed PeerConnectionEndToEndTest to use a separate worker thread.
This is a follow up to https://codereview.webrtc.org/1859933002 to change this test also to use a separate worker thread.

PeerConnectionEndToEndTest currently use the current thread both as a signaling thread and a worker thread. Although convenient while debugging, it can also hide real bugs. An example is https://codereview.webrtc.org/1766653002/#ps420001 where the worker thread is deadlocked in the track proxy due to that the worker thread waits for the signaling thread but the proxy in turns invokes the worker thread..... That bug was only discovered on Android.

BUG= webrtc:5426

Review URL: https://codereview.webrtc.org/1860423002

Cr-Commit-Position: refs/heads/master@{#12295}